Renamed "Overlay" to "Layer" Fixed Canvases not loading properly from Cues (caused by comparing a uint16_t buffer index to a uint8_t read index)

+++ src/cue/cuecontroller.cpp
@@ -184,7 +184,7 @@ namespace PixelMaestro {
 				buffer_[ID3Byte] = id_[ID3Byte];
 				read_index_ = ID3Byte;
 			}
-			else if ((uint16_t)read_index_ + 1 > buffer_size_) {
+			else if ((uint32_t)read_index_ + 1 > buffer_size_) {
 				read_index_ = 0;
 			}
 		}
